Separate ExpandableEvent to several components
